Update to clap 4

+++ src/dfx/src/main.rs
@@ -5,7 +5,7 @@ use crate::lib::environment::{Environment, EnvironmentImpl};
 use crate::lib::logger::{create_root_logger, LoggingMode};
 
 use anyhow::Error;
-use clap::{Args, Parser};
+use clap::{ArgAction, Args, Parser};
 use lib::diagnosis::{diagnose, Diagnosis, NULL_DIAGNOSIS};
 use semver::Version;
 use std::path::PathBuf;
@@ -21,15 +21,15 @@ mod util;
 #[clap(name("dfx"), version = dfx_version_str())]
 pub struct CliOpts {
     /// Displays detailed information about operations. -vv will generate a very large number of messages and can affect performance.
-    #[clap(long, short('v'), parse(from_occurrences), global(true))]
-    verbose: u64,
+    #[clap(long, short('v'), action = ArgAction::Count, global(true))]
+    verbose: u8,
 
     /// Suppresses informational messages. -qq limits to errors only; -qqqq disables them all.
-    #[clap(long, short('q'), parse(from_occurrences), global(true))]
-    quiet: u64,
+    #[clap(long, short('q'), action = ArgAction::Count, global(true))]
+    quiet: u8,
 
     /// The logging mode to use. You can log to stderr, a file, or both.
-    #[clap(long("log"), default_value("stderr"), possible_values(&["stderr", "tee", "file"]), global(true))]
+    #[clap(long("log"), default_value("stderr"), value_parser(["stderr", "tee", "file"]), global(true))]
     logmode: String,
 
     /// The file to log to, if logging to a file (see --logmode).
